2019-09-25 Version 0.1.0.0 of Poseidon
Bootstrap the MVP of Poseidon, the Simple and Extensible PostgreSQL
library.

What does it mean Simple and Extensible?

- Simple: It is a library that takes a SQL query (as text), and returns
the result as datatype. **It is not an ORM**.
- Extensible: As a use of the library, you can add your own base type to
deserialise the binary results of PostgreSQL. As of an example, if
you're working with postgis and some GeoJSON representation, you will
maybe deserialise the value into a GeospatialGeometry haskell datatype.
This example is already implemented in the `poseidon-postgis` library.

If this library is useful to you, don't hesitate to reach out to me for
new feature or update.